Add single-threaded script - rename old s3bulkdelete.py to s3threadedbulkdelete.py
